+++ crates/ryo3-bytes/src/lib.rs
@@ -22,6 +22,8 @@ pub use ::bytes::Bytes;
 pub use readable_buffer::{ExactReadableBuffer, ReadableBuffer};
 
 pub use crate::bytes::PyBytes;
+// export alias `RyBytes` to avoid confusion with `pyo3::types::PyBytes`
+pub use crate::bytes::PyBytes as RyBytes;
 
 /// ryo3-bytes python module registration
 pub fn pymod_add(m: &Bound<'_, PyModule>) -> PyResult<()> {